+++ release_notes.md
@@ -1,5 +1,9 @@
 # **Release notes**
 
+## **Release Notes develop**
+
+- When creating new sessions, the local machine and now the local lab server are checked for the most recent previous sessions
+
 ## **Release Notes 6.5.3**
 
 - Removed botched releases and patched release notes accordingly.